refactor(cli): move safeShutdownStore to lib directory for better organization
