+++ packages/cli/src/cli.ts
@@ -33,7 +33,7 @@ import { createStore } from "./livekit/store";
 import { registerMcpCommand } from "./mcp";
 import { registerModelCommand } from "./model";
 import { OutputRenderer } from "./output-renderer";
-import { safeShutdownStore } from "./shutdown";
+import { safeShutdownStore } from "./lib/shutdown";
 import { registerTaskCommand } from "./task";
 import { TaskRunner } from "./task-runner";
 import { checkForUpdates, registerUpgradeCommand } from "./upgrade";
